backend/mini_ecommerce/requirements.txt
Django==5.2.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.0
django-extensions==3.2.3
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.4.0